add yaml testutil
